# openapi.yml
openapi: 3.0.3
info:
  title: POI & Routing API
  version: 1.0.0
  description: >
    API for querying points of interest (POIs), computing routes via GraphHopper,
    and persisting public/private user routes.
servers:
  - url: /api/v1
    description: Local development
  - url: https://poi-api.dr3ff1c13nt.com/api/v1
    description: Production
paths:
...
  /pois:
    get:
      summary: List / search POIs
      description: Query the open dataset of POIs by text, category, or proximity.
      tags:
        - POIs
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Free text search (e.g. "beach", "museum").
        - in: query
          name: category
          schema:
            type: string
          description: Filter by category (e.g. "beach", "restaurant").
        - in: query
          name: lat
          schema:
            type: number
            format: double
          description: Latitude for nearby search.
        - in: query
          name: lon
          schema:
            type: number
            format: double
          description: Longitude for nearby search.
        - in: query
          name: radius
          schema:
            type: integer
            minimum: 1
          description: Radius in meters for nearby search.
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 500
          description: Maximum number of POIs to return.
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Pagination offset.
      responses:
        '200':
          description: List of POIs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoiListResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pois/{id}:
    get:
      summary: Get POI by ID
      tags:
        - POIs
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: POI identifier.
      responses:
        '200':
          description: POI details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poi'
        '404':
          description: POI not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Poi:
      type: object
      required:
        - id
        - name
        - location
      properties:
        id:
          type: string
          example: poi_1023
        name:
          type: string
          example: Paleokastritsa Beach
        category:
          type: string
          example: beach
        description:
          type: string
          nullable: true
          example: Beautiful beach with crystal clear water.
        location:
          type: object
          required:
            - lat
            - lon
          properties:
            lat:
              type: number
              format: double
              example: 39.6682
            lon:
              type: number
              format: double
              example: 19.7098
        properties:
          type: object
          additionalProperties: true
          description: Arbitrary metadata (amenities, tags, etc).
    PoiListResponse:
      type: object
      properties:
        query:
          type: object
          additionalProperties: true
          description: Echo of the query parameters.
        count:
          type: integer
          example: 3
        results:
          type: array
          items:
            $ref: '#/components/schemas/Poi'
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Missing lat/lon parameters.
        status:
          type: integer
          example: 400
        request_id:
          type: string
          nullable: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []